+++ src/main/java/com/easypost/model/ChildCollection.java
@@ -6,11 +6,11 @@ import java.util.Map;
 import lombok.Getter;
 
 @Getter
-public final class ChildrenCollection extends PaginatedCollection<Children> {
-    private List<Children> children;
+public final class ChildCollection extends PaginatedCollection<Child> {
+    private List<Child> children;
 
     @Override
-    protected Map<String, Object> buildNextPageParameters(List<Children> children, Integer pageSize) {
+    protected Map<String, Object> buildNextPageParameters(List<Child> children, Integer pageSize) {
         String lastId = children.get(children.size() - 1).getId();
 
         Map<String, Object> parameters = new java.util.HashMap<>();